changed logic of migrate method call, added MigrationController, changed MigrationTool
