+++ build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'java'
+    id 'application'
 }
 
 group = 'innowise.internship'
@@ -8,7 +9,9 @@ version = '1.0-SNAPSHOT'
 repositories {
     mavenCentral()
 }
-
+application {
+    mainClass = 'innowise.internship.Main'
+}
 dependencies {
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
