+++ src/main/java/innowise/internship/Main.java
@@ -4,6 +4,6 @@ import innowise.internship.controller.MigrationTool;
 
 public class Main {
     public static void main(String[] args) {
-        MigrationTool.run();
+        MigrationTool.run(args);
     }
 }
